+++ src/Client.php
@@ -77,8 +77,8 @@ class Client
                 ) {
                     throw new ImageNotValidException();
                 }
-                if ($this->options['max_image_size'] !== 0
-                    && $response->getHeaderLine('Content-Length') > $this->options['max_image_size']
+                if ($this->options['max_image_size'] !== 0 &&
+                    $response->getHeaderLine('Content-Length') > $this->options['max_image_size']
                 ) {
                     $size = $response->getHeaderLine('Content-Length');
                     $imageSize = Utils::formatSizeUnits($size);